deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib.pyplot
plotly
numpy
Pillow
pickle
nltk
regex
Sastrawi
scikit.learn
scipy